DROOLS-2306: Refacotr Analyzer tests to use one parent class
