+++ drools-wb-screens/drools-wb-guided-dtable-editor/drools-wb-guided-dtable-editor-client/src/test/java/org/drools/workbench/screens/guided/dtable/client/widget/analysis/BRLFragmentsAnalyzerFromFileTest.java
@@ -47,18 +47,13 @@ import static org.junit.Assert.*;
 import static org.mockito.Mockito.*;
 
 @RunWith(GwtMockitoTestRunner.class)
-public class BRLFragmentsAnalyzerFromFileTest {
+public class BRLFragmentsAnalyzerFromFileTest extends BaseDecisionTableAnalyzerTest {
 
     @GwtMock
     AnalysisConstants analysisConstants;
 
     @GwtMock
     DateTimeFormat dateTimeFormat;
-
-    @Mock
-    AsyncPackageDataModelOracle oracle;
-
-    private AnalysisReport analysisReport;
 
     @Before
     public void setUp() throws Exception {
@@ -82,44 +77,5 @@ public class BRLFragmentsAnalyzerFromFileTest {
         }
 
         assertTrue(analysisData.isEmpty());
-    }
-
-    private DecisionTableAnalyzer getDecisionTableAnalyzer(final GuidedDecisionTable52 table52) {
-        return new DecisionTableAnalyzer(mock(PlaceRequest.class),
-                                         oracle,
-                                         table52,
-                                         mock(EventBus.class)) {
-            @Override
-            protected void sendReport(AnalysisReport report) {
-                analysisReport = report;
-            }
-
-            @Override
-            protected Checks getChecks() {
-                return new Checks() {
-                    @Override
-                    protected void doRun(final CancellableRepeatingCommand command) {
-                        while (command.execute()) {
-                            //loop
-                        }
-                    }
-                };
-            }
-
-            @Override
-            protected ParameterizedCommand<Status> getOnStatusCommand() {
-                return null;
-            }
-
-            @Override
-            protected Command getOnCompletionCommand() {
-                return new Command() {
-                    @Override
-                    public void execute() {
-                        sendReport(makeAnalysisReport());
-                    }
-                };
-            }
-        };
     }
 }
