+++ drools-wb-screens/drools-wb-guided-dtable-editor/drools-wb-guided-dtable-editor-client/src/test/java/org/drools/workbench/screens/guided/dtable/client/widget/analysis/BaseDecisionTableAnalyzerTest.java
@@ -1,0 +1,76 @@
+/*
+ * Copyright 2018 Red Hat, Inc. and/or its affiliates.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.drools.workbench.screens.guided.dtable.client.widget.analysis;
+
+import com.google.gwt.event.shared.EventBus;
+import org.drools.workbench.models.guided.dtable.shared.model.GuidedDecisionTable52;
+import org.drools.workbench.screens.guided.dtable.client.widget.analysis.checks.base.Checks;
+import org.drools.workbench.screens.guided.dtable.client.widget.analysis.panel.AnalysisReport;
+import org.kie.workbench.common.widgets.client.datamodel.AsyncPackageDataModelOracle;
+import org.mockito.Mock;
+import org.uberfire.mvp.Command;
+import org.uberfire.mvp.ParameterizedCommand;
+import org.uberfire.mvp.PlaceRequest;
+
+import static org.mockito.Mockito.mock;
+
+public abstract class BaseDecisionTableAnalyzerTest {
+
+    @Mock
+    protected AsyncPackageDataModelOracle oracle;
+
+    protected AnalysisReport analysisReport;
+
+    protected DecisionTableAnalyzer getDecisionTableAnalyzer(GuidedDecisionTable52 table52) {
+        return new DecisionTableAnalyzer(mock(PlaceRequest.class),
+                                         oracle,
+                                         table52,
+                                         mock(EventBus.class)) {
+            @Override
+            protected void sendReport(AnalysisReport report) {
+                analysisReport = report;
+            }
+
+            @Override
+            protected Checks getChecks() {
+                return new Checks() {
+                    @Override
+                    protected void doRun(final CancellableRepeatingCommand command) {
+                        while (command.execute()) {
+                            //loop
+                        }
+                    }
+                };
+            }
+
+            @Override
+            protected ParameterizedCommand<Status> getOnStatusCommand() {
+                return null;
+            }
+
+            @Override
+            protected Command getOnCompletionCommand() {
+                return new Command() {
+                    @Override
+                    public void execute() {
+                        sendReport(makeAnalysisReport());
+                    }
+                };
+            }
+        };
+    }
+}
