+++ drools-wb-screens/drools-wb-guided-dtable-editor/drools-wb-guided-dtable-editor-client/src/test/java/org/drools/workbench/screens/guided/dtable/client/widget/analysis/DecisionTableAnalyzerConflictTest.java
@@ -49,17 +49,13 @@ import static org.junit.Assert.*;
 import static org.mockito.Mockito.*;
 
 @RunWith(GwtMockitoTestRunner.class)
-public class DecisionTableAnalyzerConflictTest {
+public class DecisionTableAnalyzerConflictTest extends BaseDecisionTableAnalyzerTest{
 
     @GwtMock
     AnalysisConstants analysisConstants;
 
     @GwtMock
     DateTimeFormat dateTimeFormat;
-
-    @Mock
-    AsyncPackageDataModelOracle oracle;
-    private AnalysisReport analysisReport;
 
     @Before
     public void setUp() throws Exception {
@@ -86,7 +82,7 @@ public class DecisionTableAnalyzerConflictTest {
                 .withData( new Object[][]{ { 1, "description", 100, 0, true } } )
                 .build();
 
-        DecisionTableAnalyzer analyzer = getAnalyser( table52 );
+        DecisionTableAnalyzer analyzer = getDecisionTableAnalyzer( table52 );
 
         analyzer.onValidate( new ValidateEvent( new HashMap<Coordinate, List<List<CellValue<? extends Comparable<?>>>>>() ) );
         assertTrue( analysisReport.getAnalysisData().isEmpty() );
@@ -107,7 +103,7 @@ public class DecisionTableAnalyzerConflictTest {
         table52.getData().get( 0 ).get( 2 ).setStringValue( "" );
         table52.getData().get( 0 ).get( 3 ).setStringValue( "" );
 
-        DecisionTableAnalyzer analyzer = getAnalyser( table52 );
+        DecisionTableAnalyzer analyzer = getDecisionTableAnalyzer( table52 );
 
         analyzer.onValidate( new ValidateEvent( new HashMap<Coordinate, List<List<CellValue<? extends Comparable<?>>>>>() ) );
         assertTrue( analysisReport.getAnalysisData().isEmpty() );
@@ -124,7 +120,7 @@ public class DecisionTableAnalyzerConflictTest {
                 .withData( new Object[][]{ { 1, "description", 100, 0 } } )
                 .build();
 
-        DecisionTableAnalyzer analyzer = getAnalyser( table52 );
+        DecisionTableAnalyzer analyzer = getDecisionTableAnalyzer( table52 );
 
         analyzer.onValidate( new ValidateEvent( new HashMap<Coordinate, List<List<CellValue<? extends Comparable<?>>>>>() ) );
         assertContains( "ImpossibleMatch", analysisReport );
@@ -140,7 +136,7 @@ public class DecisionTableAnalyzerConflictTest {
                 .withData( new Object[][]{ { 1, "description", "Toni", "" } } )
                 .build();
 
-        DecisionTableAnalyzer analyzer = getAnalyser( table52 );
+        DecisionTableAnalyzer analyzer = getDecisionTableAnalyzer( table52 );
 
         analyzer.onValidate( new ValidateEvent( new HashMap<Coordinate, List<List<CellValue<? extends Comparable<?>>>>>() ) );
         assertDoesNotContain( "ImpossibleMatch", analysisReport );
@@ -167,7 +163,7 @@ public class DecisionTableAnalyzerConflictTest {
                         { 2, "description", true, false, true } } )
                 .build();
 
-        DecisionTableAnalyzer analyzer = getAnalyser( table52 );
+        DecisionTableAnalyzer analyzer = getDecisionTableAnalyzer( table52 );
 
         analyzer.onValidate( new ValidateEvent( new HashMap<Coordinate, List<List<CellValue<? extends Comparable<?>>>>>() ) );
 
@@ -188,7 +184,7 @@ public class DecisionTableAnalyzerConflictTest {
                         { 2, "description", null, "true" } } )
                 .build();
 
-        DecisionTableAnalyzer analyzer = getAnalyser( table52 );
+        DecisionTableAnalyzer analyzer = getDecisionTableAnalyzer( table52 );
 
         analyzer.onValidate( new ValidateEvent( new HashMap<Coordinate, List<List<CellValue<? extends Comparable<?>>>>>() ) );
 
@@ -215,52 +211,12 @@ public class DecisionTableAnalyzerConflictTest {
                 } )
                 .build();
 
-        DecisionTableAnalyzer analyzer = getAnalyser( table52 );
+        DecisionTableAnalyzer analyzer = getDecisionTableAnalyzer( table52 );
 
         analyzer.onValidate( new ValidateEvent( new HashMap<Coordinate, List<List<CellValue<? extends Comparable<?>>>>>() ) );
 
         assertContains( "ConflictingRows", analysisReport, 2 );
         assertContains( "ConflictingRows", analysisReport, 3 );
 
-    }
-
-    private DecisionTableAnalyzer getAnalyser( final GuidedDecisionTable52 table52 ) {
-        return new DecisionTableAnalyzer( mock( PlaceRequest.class ),
-                                          oracle,
-                                          table52,
-                                          mock( EventBus.class ) ) {
-            @Override
-            protected void sendReport( AnalysisReport report ) {
-                analysisReport = report;
-            }
-
-            @Override
-            protected Checks getChecks() {
-                return new Checks() {
-                    @Override
-                    protected void doRun( final CancellableRepeatingCommand command ) {
-                        while ( command.execute() ) {
-                            //loop
-                        }
-                    }
-                };
-            }
-
-            @Override
-            protected ParameterizedCommand<Status> getOnStatusCommand() {
-                return null;
-            }
-
-            @Override
-            protected Command getOnCompletionCommand() {
-                return new Command() {
-                    @Override
-                    public void execute() {
-                        sendReport( makeAnalysisReport() );
-                    }
-                };
-            }
-
-        };
     }
 }
