+++ drools-wb-screens/drools-wb-guided-dtable-editor/drools-wb-guided-dtable-editor-client/src/test/java/org/drools/workbench/screens/guided/dtable/client/widget/analysis/DecisionTableAnalyzerDeficiencyTest.java
@@ -47,16 +47,12 @@ import static org.drools.workbench.screens.guided.dtable.client.widget.analysis.
 import static org.mockito.Mockito.*;
 
 @RunWith(GwtMockitoTestRunner.class)
-public class DecisionTableAnalyzerDeficiencyTest {
+public class DecisionTableAnalyzerDeficiencyTest extends BaseDecisionTableAnalyzerTest {
 
     @GwtMock
     AnalysisConstants analysisConstants;
     @GwtMock
     DateTimeFormat dateTimeFormat;
-
-    @Mock
-    AsyncPackageDataModelOracle oracle;
-    private AnalysisReport analysisReport;
 
     @Before
     public void setUp() throws Exception {
@@ -218,46 +214,6 @@ public class DecisionTableAnalyzerDeficiencyTest {
         assertDoesNotContain( "DeficientRow", analysisReport, 3 );
         assertDoesNotContain( "DeficientRow", analysisReport, 4 );
 
-    }
-
-    private DecisionTableAnalyzer getDecisionTableAnalyzer( GuidedDecisionTable52 table52 ) {
-        return new DecisionTableAnalyzer( mock( PlaceRequest.class ),
-                                          oracle,
-                                          table52,
-                                          mock( EventBus.class ) ) {
-            @Override
-            protected void sendReport( AnalysisReport report ) {
-                DecisionTableAnalyzerDeficiencyTest.this.analysisReport = report;
-            }
-
-            @Override
-            protected Checks getChecks() {
-                return new Checks() {
-                    @Override
-                    protected void doRun( final CancellableRepeatingCommand command ) {
-                        while ( command.execute() ) {
-                            //loop
-                        }
-                    }
-                };
-            }
-
-            @Override
-            protected ParameterizedCommand<Status> getOnStatusCommand() {
-                return null;
-            }
-
-            @Override
-            protected Command getOnCompletionCommand() {
-                return new Command() {
-                    @Override
-                    public void execute() {
-                        sendReport( makeAnalysisReport() );
-                    }
-                };
-            }
-
-        };
     }
 
 }
